Updated to massim 1.7. Note cast from double to int in RouteFinder

+++ src/mapc2017/data/RouteFinder.java
@@ -19,7 +19,7 @@ public class RouteFinder extends CityMap {
 			double minLat, double maxLat, double minLon, double maxLon,
 			Location center, int proximity) 
 	{
-		super(mapName, cellSize, minLat, maxLat, minLon, maxLon, center);
+		super(mapName, (int) cellSize, minLat, maxLat, minLon, maxLon, center);
 		
 		Location.setProximity(proximity);
